pyproject.toml
[project]
name = "defectra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "pillow-heif>=0.16.0",
    "ngrok>=1.7.0",
]
